Fix nullable results in Rx chain
RxJava 2 does not allow null values in the chain, so getting values from Map and then filtering if not null is not correct. Instead, we want to filter the stream to only keys, contained in Map, and then get corresponding values.

Also, we want to use map instead of flatMap to simply convert map's key into its result.